Update main.yml
There is a bug which need to be fixed. If the variable is present in a role then it is not overridden if we provide that in the global inventory. I fixed the crs_repos_to_enable and added  - rhel-7-server-rpms and - rhel-7-server-extras-rpms. This variable must be totally commented out or deleted in future...